refactor: 게시글 이미지비지니스 로직  리팩토링
