+++ client/src/constants/constant.ts
@@ -27,6 +27,7 @@ export const errorMessage = Object.freeze({
   blankComment: '댓글을 입력해주세요.',
   notComplete: '준비 중인 기능입니다.',
   imageCapacityExceeded: '이미지 용량 초과 입니다.',
+  failedUploadImage: '이미지 업로드에 실패했습니다. 다시 시도해주세요.',
   failedSendingEmail: '이메일 전송에 실패했습니다. 다시 시도해주세요',
   notmatchConfirmNumber:
     '메일의 인증번호와 일치하지 않습니다. 인증번호를 확인해주세요.',
