+++ client/src/service/post/index.ts
@@ -2,5 +2,12 @@ import createPost from './createPost';
 import getPost from './getPost';
 import updatePost from './updatePost';
 import deletePost from './deletePost';
+import uploadImage from './uploadImage';
 
-export const postManager = { getPost, createPost, updatePost, deletePost };
+export const postManager = {
+  getPost,
+  createPost,
+  updatePost,
+  deletePost,
+  uploadImage,
+};
